uses webpack aliases directly in postcss-import
