+++ packages/core/webpack/dev-client.js
@@ -6,9 +6,10 @@ const HardSourceWebpackPlugin = require("hard-source-webpack-plugin"),
       webpack = require("webpack"),
       yn = require("yn");
 
+const commonConfig = require("./config/common");
+
 const assetsPath = path.join(appDir, process.env.CANON_STATIC_FOLDER || "static", "assets");
 const publicPath = "/assets/";
-const appPath = path.join(appDir, "app");
 
 const loaderPath = require.resolve("./config/loaders");
 delete require.cache[loaderPath];
@@ -44,21 +45,7 @@ module.exports = {
       server: false
     })
   },
-  resolve: {
-    alias: {
-      $root: appDir,
-      $app: appPath
-    },
-    modules: [
-      appPath,
-      appDir,
-      path.resolve(__dirname, "../src"),
-      path.resolve(appDir, "node_modules"),
-      path.resolve(__dirname, "../node_modules"),
-      "node_modules"
-    ],
-    extensions: [".js", ".jsx", ".css"]
-  },
+  resolve: commonConfig.resolve,
   plugins: [
     new webpack.NormalModuleReplacementPlugin(
       /node_modules\/@datawheel\/.*\.(css|scss|sass)$/,
